+++ src/arcadia_microscopy_tools/__init__.py
@@ -1,8 +1,9 @@
+from arcadia_microscopy_tools.blending import overlay_channels
 from arcadia_microscopy_tools.channels import Channel
 from arcadia_microscopy_tools.microscopy import MicroscopyImage
 from arcadia_microscopy_tools.pipeline import ImageOperation, Pipeline, PipelineParallelized
 
-__version__ = "0.2.1"
+__version__ = "0.2.2"
 
 __all__ = [
     "Channel",
@@ -10,4 +11,5 @@ __all__ = [
     "ImageOperation",
     "Pipeline",
     "PipelineParallelized",
+    "overlay_channels",
 ]